+++ virtcontainers/utils/utils_linux.go
@@ -6,10 +6,13 @@
 package utils
 
 import (
+	"bufio"
 	"crypto/rand"
 	"fmt"
+	"io"
 	"math/big"
 	"os"
+	"strings"
 	"syscall"
 	"unsafe"
 
@@ -84,4 +87,57 @@ func FindContextID() (*os.File, uint64, error) {
 
 	vsockFd.Close()
 	return nil, 0, fmt.Errorf("Could not get a unique context ID for the vsock : %s", err)
+}
+
+const (
+	procMountsFile = "/proc/mounts"
+
+	fieldsPerLine = 6
+)
+
+const (
+	procDeviceIndex = iota
+	procPathIndex
+	procTypeIndex
+)
+
+// GetDevicePathAndFsType gets the device for the mount point and the file system type
+// of the mount.
+func GetDevicePathAndFsType(mountPoint string) (devicePath, fsType string, err error) {
+	if mountPoint == "" {
+		err = fmt.Errorf("Mount point cannot be empty")
+		return
+	}
+
+	var file *os.File
+
+	file, err = os.Open(procMountsFile)
+	if err != nil {
+		return
+	}
+
+	defer file.Close()
+
+	reader := bufio.NewReader(file)
+	for {
+		var line string
+
+		line, err = reader.ReadString('\n')
+		if err == io.EOF {
+			err = fmt.Errorf("Mount %s not found", mountPoint)
+			return
+		}
+
+		fields := strings.Fields(line)
+		if len(fields) != fieldsPerLine {
+			err = fmt.Errorf("Incorrect no of fields (expected %d, got %d)) :%s", fieldsPerLine, len(fields), line)
+			return
+		}
+
+		if mountPoint == fields[procPathIndex] {
+			devicePath = fields[procDeviceIndex]
+			fsType = fields[procTypeIndex]
+			return
+		}
+	}
 }
